mvgitlib: Remove reference branches from all_providers
The all_providers attribute contains a list list of all provider
branches (direct and indirect) for a branch.  Exclude 'reference'
providers from this list.  This also affects the branches listed
in all_dependents.

The only current user of this attribute is "git changes --dependents',
and this is the best interpretation for that usage.  If it turns out
that there are other uses, we may need to add a flag to decide
whether to exclude the 'reference' branches.

Signed-off-by: Dale Farnsworth <dfarnsworth@mvista.com>