cherry-pick-mv: Fix handling of skipped commits
git-commit (and git-commit-mv) return non-zero when there were no
changes to commit.  We now continue in that case.  Also when the
user skips a commit via "git cherry-pick-mv --skip" we include that
commit in the summary of skipped commits at the end.  Also, when
picking only one commit, don't summarize skipped commits.

Signed-off-by: Dale Farnsworth <dfarnsworth@mvista.com>