mvgitlib: Optimize the lookup of a limb's base upstream_version
Now, we calculate a limb's upstream_version only once and cache
it, re-using that value anytime the upstream_version for any branch
in that limb is requested.

This area of the limb management is still somewhat of a kludge,
since the upstream_version is maintained only in some versions
of a limb, but it's needed whenever any branch in the limb is
updated.

It is further complicated by the fact that we now support changing
the upstream_version of a limb (by rebasing the entire limb on a
new kernel version), so we must make sure we find the newer version
of upstream_version.

Signed-off-by: Dale Farnsworth <dfarnsworth@mvista.com>