Get rid of the "first commits" stuff
The "first commits" handling to try to handle branches with merge
commits was just not working.  Instead, add the concept of having
a "MONTAVISTA/merge_base" file that holds where you want to start
merging from.  If you get a repository that is not based on the
one you want, you can do a little hand merging thing add the
merge_base file to the place where you want it to start merging
from.

Signed-off-by: Corey Minyard <cminyard@mvista.com>